+++ src/main/java/com/example/picket/domain/ranking/entity/HotShow.java
@@ -10,11 +10,11 @@ import java.time.LocalDateTime;
 public class HotShow {
     private Long showId;
     private String title;
-    private Long viewCount;
+    private int viewCount;
     private String status;
     private LocalDateTime createdAt;
 
-    private HotShow(Long showId, String title, Long viewCount, String status, LocalDateTime createdAt) {
+    private HotShow(Long showId, String title, int viewCount, String status, LocalDateTime createdAt) {
         this.showId = showId;
         this.title = title;
         this.viewCount = viewCount;
@@ -22,7 +22,7 @@ public class HotShow {
         this.createdAt = createdAt;
     }
 
-    public static HotShow toEntity(Long showId, String title, Long viewCount, String status, LocalDateTime createdAt) {
+    public static HotShow toEntity(Long showId, String title, int viewCount, String status, LocalDateTime createdAt) {
         return new HotShow(showId, title, viewCount, status, createdAt);
     }
 }
